requirements.txt
Flask==2.1.2
gunicorn==20.1.0
Werkzeug==2.0.0
PyJWT==2.8.0
waitress==3.0.0
requests==2.28.1
pycryptodome==3.15.0